backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
sentence-transformers
scikit-learn
pandas
pillow
torch
transformers
